[DROOLS-4343] change FactHandle ID from a Integer into a Long

+++ drools-core/src/main/java/org/drools/core/impl/InternalKnowledgeBase.java
@@ -72,7 +72,7 @@ public interface InternalKnowledgeBase extends KieBase {
 
     FactHandleFactory newFactHandleFactory();
 
-    FactHandleFactory newFactHandleFactory(int id, long counter) throws IOException;
+    FactHandleFactory newFactHandleFactory(long id, long counter) throws IOException;
 
     Map<String, Class<?>> getGlobals();
